Fix narrowing conversion of type in RemoveCa.

+++ src/ssl.c
@@ -6143,7 +6143,7 @@ int AddCA(WOLFSSL_CERT_MANAGER* cm, DerBuffer** pDer, int type, int verify)
 }
 
 /* Removes the CA with the passed in subject hash from the cert manager's CA cert store. */
-int RemoveCA(WOLFSSL_CERT_MANAGER* cm, byte* hash, byte type)
+int RemoveCA(WOLFSSL_CERT_MANAGER* cm, byte* hash, int type)
 {
     Signer* current;
     Signer* prev;
